Move common JSON API code to helper class.
